+++ src/base/operations/sale_request_builder.js
@@ -34,7 +34,6 @@ export class SaleRequestBuilder {
     let createSaleCreationRequestOp = new xdr.CreateSaleCreationRequestOp({
       requestId: UnsignedHyper.fromString(opts.requestID),
       request: request,
-      allTasks: opts.allTasks,
       ext: new xdr.CreateSaleCreationRequestOpExt(xdr.LedgerVersion.emptyVersion())
     })
     let opAttributes = {}
@@ -136,10 +135,6 @@ export class SaleRequestBuilder {
         saleTypeExt = xdr.SaleTypeExt.fixedPrice(fixedPriceSale)
         break
       }
-    }
-
-    if (isUndefined(opts.allTasks)) {
-      opts.allTasks = 0
     }
 
     if (!BaseOperation.isValidAmount(opts.requiredBaseAssetForHardCap, true)) {
@@ -246,7 +241,6 @@ export class SaleRequestBuilder {
         asset: request.quoteAssets()[i].quoteAsset().toString()
       })
     }
-    result.allTasks = attrs.allTasks()
   }
 
   static cancelSaleCreationRequestToObject (result, attrs) {