+++ src/base/operations/sale_request_builder.spec.js
@@ -15,7 +15,6 @@ describe('SaleRequestBuilder', () => {
       endTime: '4123425',
       softCap: '20000.21',
       hardCap: '648251',
-      allTasks: 1,
       requiredBaseAssetForHardCap: '762354',
       sequenceNumber: 12,
       creatorDetails: {
@@ -48,7 +47,6 @@ describe('SaleRequestBuilder', () => {
     expect(opt.endTime).to.be.equal(obj.endTime)
     expect(opt.softCap).to.be.equal(obj.softCap)
     expect(opt.hardCap).to.be.equal(obj.hardCap)
-    expect(opt.allTasks).to.be.equal(obj.allTasks)
     expect(opt.requiredBaseAssetForHardCap).to.be.equal(obj.requiredBaseAssetForHardCap)
     expect(JSON.stringify(opt.quoteAssets)).to.be.equal(JSON.stringify(obj.quoteAssets))
     expect(isEqual(opt.details, obj.details)).to.be.true
@@ -62,7 +60,6 @@ describe('SaleRequestBuilder', () => {
       endTime: '4123425',
       softCap: '20000.21',
       hardCap: '648251',
-      allTasks: 1,
       requiredBaseAssetForHardCap: '762354',
       sequenceNumber: 13,
       creatorDetails: {
@@ -96,7 +93,6 @@ describe('SaleRequestBuilder', () => {
     expect(opt.endTime).to.be.equal(obj.endTime)
     expect(opt.softCap).to.be.equal(obj.softCap)
     expect(opt.hardCap).to.be.equal(obj.hardCap)
-    expect(opt.allTasks).to.be.equal(obj.allTasks)
     expect(opt.requiredBaseAssetForHardCap).to.be.equal(obj.requiredBaseAssetForHardCap)
     expect(opt.quoteAssets).to.be.jsonEqual(obj.quoteAssets)
     expect(isEqual(opt.details, obj.details)).to.be.true